code refactored

+++ src/main/java/org/torusresearch/torusutils/apis/KeyLookupResult.java
@@ -4,16 +4,17 @@ import org.torusresearch.torusutils.types.GetOrSetNonceResult;
 import org.torusresearch.torusutils.types.JRPCResponse;
 import org.torusresearch.torusutils.types.VerifierLookupResponse;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class KeyLookupResult {
     private final String keyResult;
-    private List<Integer> nodeIndexes;
+    private List<BigInteger> nodeIndexes;
     private final String errResult;
     private JRPCResponse<VerifierLookupResponse> errorResult;
     private GetOrSetNonceResult nonceResult;
 
-    public KeyLookupResult(String _keyResult, String _errResult, List<Integer> nodeIndexes, GetOrSetNonceResult nonceResult) {
+    public KeyLookupResult(String _keyResult, String _errResult, List<BigInteger> nodeIndexes, GetOrSetNonceResult nonceResult) {
         keyResult = _keyResult;
         errResult = _errResult;
         this.nodeIndexes = nodeIndexes;
@@ -25,7 +26,7 @@ public class KeyLookupResult {
         errResult = _errResult;
     }
 
-    public List<Integer> getNodeIndexes() {
+    public List<BigInteger> getNodeIndexes() {
         return nodeIndexes;
     }
